+++ lib/cheftacular/helper.rb
@@ -28,10 +28,16 @@ class Cheftacular
       !@config['action'].public_methods(false).include?(command.to_sym) && !@config['stateless_action'].public_methods(false).include?(command.to_sym)
     end
 
-    def running_on_chef_node?
-      Dir.entries('/etc').include?('chef')
+    def is_initialization_command? command=''
+      command ||= ''
+            
+      @config['initialization_action'].public_methods(false).include?(command.to_sym)
+    end
+
+    def running_on_chef_node? ret = false
+      Dir.entries('/etc').include?('chef') && File.exist?('/etc/chef/client.rb') && !File.size?('/etc/chef/client.rb').nil?
     rescue StandardError => e
-      exception_output "An error occurred while trying to see if this system is a chef node. Assuming the system is not a chef node.", e, false
+      @config['error'].exception_output "An error occurred while trying to see if this system is a chef node. Assuming the system is not a chef node.", e, false
     end
 
     def running_in_mode? mode
@@ -42,10 +48,6 @@ class Cheftacular
       puts "Checking remote #{ declassify } version..."
 
       `gem list #{ declassify } --remote`[/(\d+\.\d+\.\d+)/]
-    end
-
-    def is_junk_filename? filename
-      filename =~ /.DS_Store|.com.apple.timemachine.supported|README.*/ || filename == '.' || filename == '..' && File.directory?(filename)
     end
 
     def completion_rate? percent, mode
@@ -102,10 +104,6 @@ class Cheftacular
       puts("\nDone in #{ Time.now - @config['start_time'] } seconds at #{ Time.now.strftime('%Y-%m-%d %l:%M:%S %P') }.") unless @options['quiet']
     end
 
-    def write_version_file version
-      File.open( current_version_file_path, "w") { |f| f.write(version) }
-    end
-
     def set_local_instance_vars
       [ 
         @options, 
@@ -117,61 +115,6 @@ class Cheftacular
         @config['cheftacular'],
         @config['server_passwords']
       ]
-    end
-
-    def exception_output message, exception='', exit_on_call=true, suppress_error_output=false
-      puts "#{ message }\n"
-
-      puts("Error message: #{ exception }\n#{ exception.backtrace.join("\n") }") unless suppress_error_output
-
-      exit if exit_on_call
-    end
-
-    def write_nodes_file_cache nodes
-      nodes.each do |node|
-        File.open( File.join( current_nodes_file_cache_path, "#{ node.name }.json"), "w") { |f| f.write(node.to_json) }
-      end
-    end
-
-    def check_nodes_file_cache nodes=[]
-      Dir.entries(current_nodes_file_cache_path).each do |location|
-        next if is_junk_filename?(location)
-
-        nodes << @config['ridley'].node.from_file("#{ current_nodes_file_cache_path }/#{ location }" )
-      end
-
-      nodes
-    end
-
-    def current_version_file_path
-      current_file_path 'version-check.txt'
-    end
-
-    def current_audit_file_path
-      current_file_path 'audit-check.txt'
-    end
-
-    def compare_file_node_cache_against_chef_nodes mode='include?'
-      chef_server_names, nodes_file_cache_names = [],[]
-      included = true 
-
-      @config['chef_nodes'].each { |node| chef_server_names << node.name }
-
-      check_nodes_file_cache.each { |node| nodes_file_cache_names << node.name }
-
-      nodes_file_cache_names.each do |node_name|
-        unless chef_server_names.include?(node_name)
-          included = false
-
-          break
-        end
-      end
-
-      case mode
-      when 'include?'  then return included
-      when 'not_equal' then return check_nodes_file_cache.count != names.count
-      when 'equal'     then return chef_server_names.sort == nodes_file_cache_names.sort
-      end
     end
 
     def is_higher_version? vstr1, vstr2
@@ -228,45 +171,6 @@ class Cheftacular
       @options['preferred_cloud_region'] = @options['preferred_cloud_region'].nil? ? @config['cheftacular']['preferred_cloud_region'] : @options['preferred_cloud_region']
       @options['virtualization_mode']    = @options['virtualization_mode'].nil? ?    @config['cheftacular']['virtualization_mode'] :    @options['virtualization_mode']
       @options['route_dns_changes_via']  = @options['route_dns_changes_via'].nil? ?  @config['cheftacular']['route_dns_changes_via'] :  @options['route_dns_changes_via']
-    end
-
-    def current_nodes_file_cache_path
-      current_file_path 'node_cache'
-    end
-
-    def cleanup_file_caches mode='old', check_current_day_entry=false
-      base_dir = File.join( @config['locs']['app-root'], 'tmp', declassify )
-
-      Dir.entries(base_dir).each do |entry|
-        next if is_junk_filename?(entry)
-
-        case mode
-        when 'old'
-          FileUtils.rm("#{ base_dir }/#{ entry }") if File.file?("#{ base_dir }/#{ entry }") && !entry.include?(Time.now.strftime("%Y%m%d"))
-        when 'current'
-          check_current_day_entry = true
-        when 'current-audit-only'
-          FileUtils.rm("#{ base_dir }/#{ entry }") if File.file?("#{ base_dir }/#{ entry }") && entry.include?(Time.now.strftime("%Y%m%d"))
-        end
-
-        if File.exists?("#{ base_dir }/#{ entry }") && File.directory?("#{ base_dir }/#{ entry }")
-          FileUtils.rm_rf("#{ base_dir }/#{ entry }") if !check_current_day_entry && !entry.include?(Time.now.strftime("%Y%m%d"))
-          
-          FileUtils.rm_rf("#{ base_dir }/#{ entry }") if check_current_day_entry && entry.include?(Time.now.strftime("%Y%m%d"))
-
-          FileUtils.mkdir_p @config['helper'].current_nodes_file_cache_path
-        end
-      end
-    end
-
-    def remove_current_file_node_cache
-      base_dir = File.join( @config['locs']['app-root'], 'tmp', declassify )
-
-      Dir.entries(base_dir).each do |entry|
-        next if is_junk_filename?(entry)
-
-        FileUtils.rm_rf("#{ base_dir }/#{ entry }") if File.directory?("#{ base_dir }/#{ entry }") && entry.include?(Time.now.strftime("%Y%m%d"))
-      end      
     end
 
     def does_cheftacular_config_have? key_array
@@ -329,14 +233,6 @@ class Cheftacular
       exit
     end
 
-    def current_chef_repo_cheftacular_file_cache_path
-      current_file_path "chef_repo_cheftacular_cache"
-    end
-
-    def write_chef_repo_cheftacular_cache_file hash
-      File.open( current_chef_repo_cheftacular_file_cache_path, "w") { |f| f.write(hash) }
-    end
-
     def compile_chef_repo_cheftacular_yml_as_hash
       master_hash = get_cheftacular_yml_as_hash
       master_hash['replace_keys_in_chef_repo'].each_pair do |key, val|
@@ -344,28 +240,6 @@ class Cheftacular
       end
 
       master_hash
-    end
-
-    def write_chef_repo_cheftacular_yml_file file_location
-      File.open( file_location, "w") { |f| f.write(compile_chef_repo_cheftacular_yml_as_hash.to_yaml) }
-    end
-
-    def is_initialization_command? command=''
-      command ||= ''
-            
-      @config['initialization_action'].public_methods(false).include?(command.to_sym)
-    end
-
-    def get_current_real_node_name other_node_name='', ret=''
-      ret << @options['env'] + @config['cheftacular']['node_name_separator']
-      ret << other_node_name.blank? ? @options['node_name'] : other_node_name
-
-      ret
-    end
-
-    private
-    def current_file_path file_name
-      File.join( @config['locs']['app-root'], 'tmp', declassify, "#{ Time.now.strftime("%Y%m%d") }-#{ file_name }")
     end
   end
 end