+++ lib/cheftacular/stateless_actions/initialize_cheftacular_yml.rb
@@ -4,7 +4,7 @@ class Cheftacular
       @config['documentation']['stateless_action'] <<  [
         "`cft initialize_cheftacular_yml` will create a cheftacular.yml file in your config folder (and create the" +
         "config folder if it does not exist). If you already have a cheftacular.yml file in the config folder, it will " +
-        "create a cheftacular.yml.example file that will contain the new changes / keys in the latest cheftacular version."
+        "create a cheftacular.example.yml file that will contain the new changes / keys in the latest cheftacular version."
       ]
     end
   end
@@ -17,7 +17,13 @@ class Cheftacular
 
   class StatelessAction
     def initialize_cheftacular_yml
-      #TODO
+      FileUtils.mkdir_p(File.join(@config['locs']['chef-repo'], "config"))
+
+      if File.exist?(File.join(@config['locs']['chef-repo'], "config", "cheftacular.yml"))
+        @config['helper'].write_config_cheftacular_yml_file('cheftacular.example.yml')
+      else
+        @config['helper'].write_config_cheftacular_yml_file
+      end
     end
   end
 end