+++ lib/cloud_interactor/sshkey.rb
@@ -32,6 +32,12 @@ class CloudInteractor
       @main_obj["#{ IDENTITY }_create_request"] = JSON.parse(@classes['auth'].auth_service(RESOURCE).instance_eval(IDENTITY).create(args).to_json)
     end
 
+    def destroy args
+      read args, false
+
+      @classes['helper'].generic_destroy_parse args, IDENTITY, RESOURCE
+    end
+
     #special method for digitalocean api, will attempt to create the sshkeyid on DO if one matching the user's hostname is not found
     def bootstrap output=true
       read Socket.gethostname, false
@@ -39,8 +45,8 @@ class CloudInteractor
       if @main_obj["specific_#{ IDENTITY }"].empty?
         puts "Did not detect an SSHKey on DigitalOcean for the system #{ Socket.gethostname }, creating..."
         create_hash = {}
-        create_hash[:name]        = Socket.gethostname
-        create_hash[:ssh_pub_key] = File.read(File.expand_path('~/.ssh/id_rsa.pub'))
+        create_hash['name']       = Socket.gethostname
+        create_hash['ssh_pub_key'] = File.read(File.expand_path('~/.ssh/id_rsa.pub'))
 
         create create_hash